habr_final/backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
email-validator==2.2.0
SQLAlchemy==2.0.36
pyodbc==5.1.0
python-dotenv==1.0.1
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.17
cloudinary==1.44.1
alembic==1.14.0
